Make an asset server app
Instead of trying to bundle the assets into the game, just make an asset server that serves the apps at a url, then the game can use a base URL to get the asset. 

Could even expose an asset endpoint to get the meta data from it so the game never needs to think about it!